backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
apscheduler>=3.10.0
openai>=1.109.1
python-dotenv>=0.19.0
requests>=2.28.0
elevenlabs
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dateutil>=2.8.0
pandas
faker
selenium
crawl4ai